add day of work

+++ dao/mariaDB.go
@@ -35,14 +35,14 @@ func AddEmployee(emp model.Employee, db *gorm.DB) model.Employee {
 
 }
 
-func GetEmployeeById(id int, db *gorm.DB) model.Employee {
+func GetEmployeeById(id uint64, db *gorm.DB) model.Employee {
 
 	emp := model.Employee{}
 	db.Where("id = " + fmt.Sprint(id)).First(&emp)
 	return emp
 }
 
-func DeleteEmployeeById(id int, db *gorm.DB) {
+func DeleteEmployeeById(id uint64, db *gorm.DB) {
 	emp := model.Employee{ID: id}
 	db.Delete(&emp)
 }
